us06 completa y en funcionamiento

+++ Spring1/src/main/java/com/Bootcamp/Spring1/dto/response/FlyResponseDTO.java
@@ -7,5 +7,5 @@ import lombok.*;
 @Data
 public class FlyResponseDTO {
     public String mensaje = "El monto total del vuelo es de: ";
-    private Double total;
+    private Integer total;
 }
